reset-pwd initial commit
