+++ app/system-modules/reset-pwd/index.js
@@ -1,0 +1,1 @@
+require('./lib');
