+++ app/system-modules/reset-pwd/lib/index.js
@@ -1,0 +1,1 @@
+require('./routes');
