+++ app/system-modules/reset-pwd/lib/routes/index.js
@@ -1,0 +1,1 @@
+require('./reset');
